Ignore exceptions when calling shutdown on sockets...
On Linux, if you call shutdown on a socket that has been closed by
the other end, it just returns normally.
On Mac OS X, it throws an exception.
The exception is never caught, so mapper.main terminates without
joining all its threads and closing all its connections.
The best bet is just to swallow the exception.